move getCategoryCount to DrawableDB, keep selection in HashTree
revert unrelated logging change
