+++ ImageGallery/src/org/sleuthkit/autopsy/imagegallery/datamodel/CategoryManager.java
@@ -1,7 +1,7 @@
 /*
  * Autopsy Forensic Browser
  *
- * Copyright 2015 Basis Technology Corp.
+ * Copyright 2015-16 Basis Technology Corp.
  * Contact: carrier <at> sleuthkit <dot> org
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,23 +23,18 @@ import com.google.common.cache.LoadingCache;
 import com.google.common.eventbus.AsyncEventBus;
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.LongAdder;
 import java.util.logging.Level;
 import javax.annotation.concurrent.Immutable;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.concurrent.BasicThreadFactory;
-import org.openide.util.Exceptions;
 import org.sleuthkit.autopsy.casemodule.events.ContentTagAddedEvent;
 import org.sleuthkit.autopsy.casemodule.events.ContentTagDeletedEvent;
 import org.sleuthkit.autopsy.coreutils.Logger;
 import org.sleuthkit.autopsy.imagegallery.ImageGalleryController;
 import org.sleuthkit.datamodel.ContentTag;
-import org.sleuthkit.datamodel.SleuthkitCase;
 import org.sleuthkit.datamodel.TagName;
 import org.sleuthkit.datamodel.TskCoreException;
 
@@ -48,8 +43,8 @@ import org.sleuthkit.datamodel.TskCoreException;
  * {@link CategoryChangeEvent}s when files are categorized.
  *
  * To receive CategoryChangeEvents, a listener must register itself, and
- * implement a method annotated with {@link Subscribe} that accepts one argument
- * of type CategoryChangeEvent
+ * implement a public method annotated with {@link Subscribe} that accepts one
+ * argument of type CategoryChangeEvent
  *
  * TODO: currently these two functions (cached counts and events) are separate
  * although they are related. Can they be integrated more?
@@ -120,31 +115,7 @@ public class CategoryManager {
         fireChange(Collections.emptyList(), null);
     }
 
-    public long getCategoryCount(Category cat, Collection<Long> fileIDs) {
-        String name = "select count(obj_id) from tsk_files "
-                + " where "
-                + "         obj_id in (" + StringUtils.join(fileIDs, ",") + " ) "
-                + "     and obj_id not in ( "
-                + "         select obj_id from  content_tags where "
-                + "             content_tags.tag_name_id in ("
-                + getTagName(Category.FIVE).getId() + " ,"
-                + getTagName(Category.FOUR).getId() + " ,"
-                + getTagName(Category.THREE).getId() + " ,"
-                + getTagName(Category.TWO).getId() + " ,"
-                + getTagName(Category.ONE).getId()
-                + " ))";
-        try (SleuthkitCase.CaseDbQuery executeQuery = controller.getSleuthKitCase().executeQuery(name);
-                ResultSet resultSet = executeQuery.getResultSet();) {
-            while (resultSet.next()) {
-                return resultSet.getLong("count(obj_id)");
-            }
-        } catch (SQLException sQLException) {
-        } catch (TskCoreException ex) {
-            Exceptions.printStackTrace(ex);
-
-        }
-        return -1;
-    }
+  
 
     /**
      * get the number of file with the given {@link Category}
